New serial port interface
To send large blocks of data (like full swing telemetry) it is helpful to be
able to point at a block of memory and copy it to the serial port. The DMASerial
object supports an enqueue method which accepts a buffer, length, and function
to be called on completion of transmission. The buffer is not copied, so large
objects can be sent without running out of memory. The completion object can be
used to decide when it is safe to re-use the buffer.

In addition, this serial port supports the usual buffered IO. When a buffered
method call is made, contiguous space is allocated in the ring buffer and passed
to the enqueue method. This may require splitting the message into multiple
sends.

When sending over the XBee radio, if CTS is high and all buffers are full, all
methods can fail to send data. If CTS is low, the methods will wait until there
is buffer space or CTS goes high.